RABSW-685: Add ClientMount resource and clientmountd daemon
This commit adds a new resource, ClientMount, that is used to set up mount points
on clients. The resource is used by a controller that runs as a daemon (outside of
kubernetes) on clients. This daemon is named clientmountd, and it is based on the
same reconcile logic as controllers within the kubernetes cluster. The clientmountd
daemon runs mount/umount commands on the host OS to create the mounts specified in
the ClientMount resource.

There is a new make target, "build-daemon", that compiles the clientmountd binary.