RABSW-883: Use owner labels to track child resources
This commit adds a set of labels to track the kind, name, and namespace
of the owner of a resource. This allows parent resource to find their
children without having to keep ObjectReferences to them. It allows deletion
of the child resources using the DeleteAllOf() call with a filter by
label.

Explicitly deleting child resources rather than relying on garbage collection
allows the parent resource to delay its deletion until the child resources are
fully deleted. This is the case regardless of whether foreground or background
deletion is used.

All child resources are also labeled with a "workflow" or "PersistentStorageInstance"
name and namespace. These labels are inherited throughout the resource chain and
allow an easy way to trace a resource back to its workflow or PersistentStorageInstance.

There are several changes to allow DWS and SOS to use this functionality:
 - The workflow controller adds owner labels to child resources it creates
 - The workflow controller uses a finalizer to prevent its deletion. Before
   the finalizer is removed, the workflow controller deletes all children.
 - The workflow webhook was changed to allow adding the finalizer before the
   status.state field is set.
 - A new interface "ObjectList" provides a GetObjectList() function for *List
   structs (e.g., ComputesList) to return a list of client.Objects rather than
   resource specific structs (e.g., Computes).
 - The Servers field in the PersistentStorageInstance resource was moved to the
   status section since the Servers resource is a child of the PersistentStorageInstance

Signed-off-by: Matt Richerson <mattr@cray.com>